refactor: reorganize server modules
