+++ apps/server/src/ClientRegistry.ts
