+++ apps/server/src/context.ts
@@ -2,8 +2,8 @@ import type { Branded } from "@mp/std";
 import type { AuthToken, AuthServer } from "@mp/auth-server";
 import type { Logger } from "@mp/logger";
 import type { StateAccess } from "@mp/sync/server";
-import type { WorldState } from "./modules/world/schema";
-import type { ClientRegistry } from "./modules/world/ClientRegistry";
+import type { WorldState } from "./WorldState";
+import type { ClientRegistry } from "./ClientRegistry";
 
 export interface ServerContext {
   sessionId: HttpSessionId;
