+++ apps/server/src/modules/character/schema.ts
@@ -1,8 +1,7 @@
 import type { AreaId } from "@mp/data";
 import { integer, pgTable, serial } from "drizzle-orm/pg-core";
 import type { Path } from "@mp/math";
-import type { UserId, UserIdentity } from "@mp/auth-server";
-import type { SyncServer } from "@mp/sync/server";
+import type { UserId } from "@mp/auth-server";
 import { branded } from "../../db/types/branded";
 import { vector } from "../../db/types/vector";
 
@@ -20,10 +19,4 @@ export interface Character extends DBCharacter {
   path?: Path;
 }
 
-export type WorldState = {
-  characters: Record<CharacterId, Character>;
-};
-
 export type CharacterId = Character["id"];
-
-export type WorldSyncServer = SyncServer<WorldState, WorldState, UserIdentity>;
