+++ apps/server/src/modules/router.ts
@@ -3,13 +3,13 @@ import { t } from "../trpc";
 import { createAreaRouter } from "./area/router";
 import type { SystemRouterDependencies } from "./system/router";
 import { createSystemRouter } from "./system/router";
-import type { WorldRouterDependencies } from "./world/router";
-import { createWorldRouter } from "./world/router";
+import type { CharacterRouterDependencies } from "./character/router";
+import { createCharacterRouter } from "./character/router";
 
 export type RootRouter = ReturnType<typeof createRootRouter>;
 
 export interface RootRouterDependencies
-  extends WorldRouterDependencies,
+  extends CharacterRouterDependencies,
     SystemRouterDependencies {
   createUrl: UrlFactory;
 }
@@ -19,7 +19,7 @@ export function createRootRouter({
   ...dependencies
 }: RootRouterDependencies) {
   return t.router({
-    world: createWorldRouter(dependencies),
+    world: createCharacterRouter(dependencies),
     system: createSystemRouter(dependencies),
     area: createAreaRouter(createUrl),
   });
