+++ apps/server/src/package.ts
@@ -1,9 +1,10 @@
 // This is the file that represents what the @mp/server package exports.
 // This module should never export runtime code. Only use type exports.
-export type * from "./modules/world/schema";
+export type * from "./modules/character/schema";
+export type * from "./WorldState";
 export type * from "./context";
 export type * from "./modules/router";
-export type { ServerOptions } from "./schemas/serverOptions";
+export type { ServerOptions } from "./serverOptions";
 
 // An exception is made for the shared code between client & server
 export * from "./shared";
